feature: move SmartDocuments URL building to DocumentCreationService
