+++ src/test/kotlin/net/atos/zac/configuratie/ConfiguratieServiceTest.kt
@@ -38,14 +38,6 @@ class ConfiguratieServiceTest : BehaviorSpec({
         val gemeenteMail = "gemeente@example.com"
         val bpmnSupport = false
         val catalogusUri = "https://example.com/catalogus"
-        val zaakUuid = UUID.randomUUID()
-        val templateGroupId = "groupId"
-        val templateId = "templateId"
-        val title = "title"
-        val description = "description"
-        val creationDate = ZonedDateTime.of(2024, 10, 7, 0, 0, 0, 0, ZoneOffset.UTC)
-        val userName = "Full User Name"
-
         every { catalogus.url } returns URI(catalogusUri)
         every { ztcClientService.readCatalogus(any<CatalogusListParameters>()) } returns catalogus
 
@@ -61,94 +53,30 @@ class ConfiguratieServiceTest : BehaviorSpec({
             bpmnSupport
         )
 
-        When("Document creation URL is requested for zaak") {
-            val uri = configurationService.documentCreationCallbackUrl(
-                zaakUuid = zaakUuid,
-                null,
-                templateGroupId,
-                templateId,
-                title,
-                description,
-                creationDate,
-                userName
-            )
+        When("zaak tonen URL is requested") {
+            val zaakTonenUrl = configurationService.zaakTonenUrl(zaakIdentificatie = "id")
 
-            Then("Correct URl is provided") {
-                uri.toString() shouldBe "$contextUrl/rest/document-creation/smartdocuments/callback/zaak/$zaakUuid" +
-                    "?templateId=$templateId" +
-                    "&templateGroupId=$templateGroupId" +
-                    "&title=$title" +
-                    "&userName=Full+User+Name" +
-                    "&creationDate=2024-10-07T00%3A00%3A00Z" +
-                    "&description=$description"
+            Then("correct url is built") {
+                zaakTonenUrl.toString() shouldBe "$contextUrl/zaken/id"
             }
         }
 
-        When("Document creation URL is requested for taak") {
-            val taakUuid = UUID.randomUUID().toString()
-            val uri = configurationService.documentCreationCallbackUrl(
-                zaakUuid,
-                taakUuid,
-                templateGroupId,
-                templateId,
-                title,
-                description,
-                creationDate,
-                userName
-            )
+        When("taak tonen URL is requested") {
+            val taakTonenUrl = configurationService.taakTonenUrl(taakId = "id")
 
-            Then("Correct URl is provided") {
-                uri.toString() shouldBe
-                    "$contextUrl/rest/document-creation/smartdocuments/callback/zaak/$zaakUuid/task/$taakUuid" +
-                    "?templateId=$templateId" +
-                    "&templateGroupId=$templateGroupId" +
-                    "&title=$title" +
-                    "&userName=Full+User+Name" +
-                    "&creationDate=2024-10-07T00%3A00%3A00Z" +
-                    "&description=$description"
+            Then("correct url is built") {
+                taakTonenUrl.toString() shouldBe "$contextUrl/taken/id"
+            }
+        }
+
+        When("informatieobject tonen url is requested") {
+            val uuid = UUID.randomUUID()
+            val informatieobjectTonenUrl = configurationService.informatieobjectTonenUrl(uuid)
+
+            Then("Correct url is built") {
+                informatieobjectTonenUrl.toString() shouldBe "$contextUrl/informatie-objecten/$uuid"
             }
         }
     }
 
-    Given("SmartDocuments wizard finished execution") {
-        val entityManager = mockk<EntityManager>()
-        val ztcClientService = mockk<ZtcClientService>()
-        val catalogus = mockk<Catalogus>()
-
-        val additionalAllowedFileTypes = ""
-        val zgwApiClientMpRestUrl = "https://example.com:1111"
-        val contextUrl = "https://example.com:2222"
-        val gemeenteCode = "gemeenteCode"
-        val gemeenteNaam = "Gemeente Name"
-        val gemeenteMail = "gemeente@example.com"
-        val bpmnSupport = false
-        val catalogusUri = "https://example.com/catalogus"
-
-        every { catalogus.url } returns URI(catalogusUri)
-        every { ztcClientService.readCatalogus(any<CatalogusListParameters>()) } returns catalogus
-
-        val configurationService = ConfiguratieService(
-            entityManager,
-            ztcClientService,
-            additionalAllowedFileTypes,
-            zgwApiClientMpRestUrl,
-            contextUrl,
-            gemeenteCode,
-            gemeenteNaam,
-            gemeenteMail,
-            bpmnSupport
-        )
-
-        When("SmartDocuments finish page URL is requested") {
-            val finishPageUrl = configurationService.documentCreationFinishPageUrl("1", "1", "document name", "result")
-
-            Then("correct URL is built") {
-                finishPageUrl.toString() shouldBe "$contextUrl/static/smart-documents-result.html" +
-                    "?zaak=1" +
-                    "&taak=1" +
-                    "&doc=document+name" +
-                    "&result=result"
-            }
-        }
-    }
 })
