+++ src/test/kotlin/net/atos/zac/documentcreation/DocumentCreationServiceTest.kt
@@ -13,10 +13,14 @@ import io.mockk.mockk
 import io.mockk.mockkStatic
 import io.mockk.verify
 import jakarta.enterprise.inject.Instance
+import jakarta.persistence.EntityManager
 import net.atos.client.smartdocuments.model.createFile
 import net.atos.client.zgw.drc.model.createEnkelvoudigInformatieObjectCreateLockRequest
 import net.atos.client.zgw.zrc.model.createZaak
 import net.atos.client.zgw.zrc.model.createZaakInformatieobject
+import net.atos.client.zgw.ztc.ZtcClientService
+import net.atos.client.zgw.ztc.model.CatalogusListParameters
+import net.atos.client.zgw.ztc.model.generated.Catalogus
 import net.atos.zac.app.informatieobjecten.EnkelvoudigInformatieObjectUpdateService
 import net.atos.zac.authentication.LoggedInUser
 import net.atos.zac.authentication.createLoggedInUser
@@ -30,6 +34,7 @@ import net.atos.zac.identity.model.getFullName
 import net.atos.zac.smartdocuments.SmartDocumentsService
 import net.atos.zac.smartdocuments.SmartDocumentsTemplatesService
 import java.net.URI
+import java.time.ZoneOffset
 import java.time.ZonedDateTime
 import java.util.UUID
 
@@ -63,6 +68,7 @@ class DocumentCreationServiceTest : BehaviorSpec({
         val data = createData()
         val documentCreationAttendedResponse = createDocumentCreationAttendedResponse()
         val fullName = "Full Name"
+        val contextUrl = "https://example.com"
 
         every { loggedInUserInstance.get() } returns loggedInUser
         mockkStatic(User::getFullName)
@@ -81,18 +87,7 @@ class DocumentCreationServiceTest : BehaviorSpec({
         every {
             smartDocumentsTemplatesService.getTemplateName(documentCreationData.templateId)
         } returns UUID.randomUUID().toString()
-        every {
-            configuratieService.documentCreationCallbackUrl(
-                documentCreationData.zaak.uuid,
-                documentCreationData.taskId,
-                documentCreationData.templateGroupId,
-                documentCreationData.templateId,
-                documentCreationData.title,
-                documentCreationData.description,
-                documentCreationData.creationDate,
-                fullName
-            )
-        } returns URI("https://sd.host.com")
+        every { configuratieService.readContextUrl() } returns contextUrl
 
         When("the 'create document attended' method is called") {
             val documentCreationResponse = documentCreationService.createDocumentAttended(documentCreationData)
@@ -171,6 +166,86 @@ class DocumentCreationServiceTest : BehaviorSpec({
                         skipPolicyCheck = true
                     )
                 }
+            }
+        }
+    }
+
+    Given("A zaak exists") {
+        val contextUrl = "https://example.com:2222"
+        val zaakUuid = UUID.randomUUID()
+        val templateGroupId = "groupId"
+        val templateId = "templateId"
+        val title = "title"
+        val description = "description"
+        val creationDate = ZonedDateTime.of(2024, 10, 7, 0, 0, 0, 0, ZoneOffset.UTC)
+        val userName = "Full User Name"
+
+        every { configuratieService.readContextUrl() } returns contextUrl
+
+        When("Document creation URL is requested for zaak") {
+            val uri = documentCreationService.documentCreationCallbackUrl(
+                zaakUuid = zaakUuid,
+                null,
+                templateGroupId,
+                templateId,
+                title,
+                description,
+                creationDate,
+                userName
+            )
+
+            Then("Correct URl is provided") {
+                uri.toString() shouldBe "$contextUrl/rest/document-creation/smartdocuments/callback/zaak/$zaakUuid" +
+                        "?templateId=$templateId" +
+                        "&templateGroupId=$templateGroupId" +
+                        "&title=$title" +
+                        "&userName=Full+User+Name" +
+                        "&creationDate=2024-10-07T00%3A00%3A00Z" +
+                        "&description=$description"
+            }
+        }
+
+        When("Document creation URL is requested for taak") {
+            val taakUuid = UUID.randomUUID().toString()
+            val uri = documentCreationService.documentCreationCallbackUrl(
+                zaakUuid,
+                taakUuid,
+                templateGroupId,
+                templateId,
+                title,
+                description,
+                creationDate,
+                userName
+            )
+
+            Then("Correct URl is provided") {
+                uri.toString() shouldBe
+                        "$contextUrl/rest/document-creation/smartdocuments/callback/zaak/$zaakUuid/task/$taakUuid" +
+                        "?templateId=$templateId" +
+                        "&templateGroupId=$templateGroupId" +
+                        "&title=$title" +
+                        "&userName=Full+User+Name" +
+                        "&creationDate=2024-10-07T00%3A00%3A00Z" +
+                        "&description=$description"
+            }
+        }
+    }
+
+    Given("SmartDocuments wizard finished execution") {
+        val contextUrl = "https://example.com"
+        every { configuratieService.readContextUrl() } returns contextUrl
+
+        When("SmartDocuments finish page URL is requested") {
+            val finishPageUrl = documentCreationService.documentCreationFinishPageUrl(
+                "1", "1", "document name", "result"
+            )
+
+            Then("correct URL is built") {
+                finishPageUrl.toString() shouldBe "$contextUrl/static/smart-documents-result.html" +
+                        "?zaak=1" +
+                        "&taak=1" +
+                        "&doc=document+name" +
+                        "&result=result"
             }
         }
     }
